+++ src/consensus/activation.h
@@ -45,6 +45,8 @@ bool IsGluonEnabled(const Consensus::Params &params,
 
 /** Check if May 15th, 2023 protocol upgrade has activated. */
 bool IsWellingtonEnabled(const Consensus::Params &params,
+                         int64_t nMedianTimePast);
+bool IsWellingtonEnabled(const Consensus::Params &params,
                          const CBlockIndex *pindexPrev);
 
 #endif // BITCOIN_CONSENSUS_ACTIVATION_H