+++ src/policy/policy.cpp
@@ -68,7 +68,9 @@ bool IsStandard(const CScript &scriptPubKey, TxoutType &whichType) {
 
 bool IsStandardTx(const CTransaction &tx, bool permit_bare_multisig,
                   const CFeeRate &dust_relay_fee, std::string &reason) {
-    if (tx.nVersion > CTransaction::MAX_STANDARD_VERSION || tx.nVersion < 1) {
+    // Only allow these tx versions, will become consensus rule in Wellington
+    if (tx.nVersion > CTransaction::MAX_STANDARD_VERSION ||
+        tx.nVersion < CTransaction::MIN_STANDARD_VERSION) {
         reason = "version";
         return false;
     }