+++ src/primitives/transaction.h
@@ -194,13 +194,15 @@ inline void SerializeTransaction(const TxType &tx, Stream &s) {
 class CTransaction {
 public:
     // Default transaction version.
-    static const int32_t CURRENT_VERSION = 2;
+    static constexpr int32_t CURRENT_VERSION = 2;
 
-    // Changing the default transaction version requires a two step process:
-    // first adapting relay policy by bumping MAX_STANDARD_VERSION, and then
-    // later date bumping the default CURRENT_VERSION at which point both
-    // CURRENT_VERSION and MAX_STANDARD_VERSION will be equal.
-    static const int32_t MAX_STANDARD_VERSION = 2;
+    // Policy: Valid min/max for nVersion.
+    // Remove after wellington activation.
+    static constexpr int32_t MIN_STANDARD_VERSION = 1, MAX_STANDARD_VERSION = 2;
+
+    // Consensus: Valid min/max for nVersion, enforced after Wellington.
+    static constexpr int32_t MIN_CONSENSUS_VERSION = 1,
+                             MAX_CONSENSUS_VERSION = 2;
 
     // The local variables are made const to prevent unintended modification
     // without updating the cached hash value. However, CTransaction is not