+++ doc/release-notes.md
@@ -13,3 +13,14 @@ This release includes the following features and fixes:
     `listunspent` RPC have been deprecated and will be removed in a future
     version. To keep using these fields, use the
     `-deprecatedrpc=mempool_ancestors_descendants` option.
+
+Network upgrade
+---------------
+
+At the MTP time of `1684152000` (May 15, 2023 12:00:00 UTC), the following
+changes will become activated:
+  - New consensus rule: The `nVersion` field in `CTransaction` now must be either
+    1 or 2. This has been a policy rule (i.e. wallets already cannot use anything
+    other than 1 or 2), but miners were still able to mine transactions with
+    versions other than 1 and 2. Disallowing them by consensus allows us to use
+    the version field for e.g. a new & scalable transaction format.